Manual User Struct Packing

+++ contracts/interfaces/ISafeTokenLock.sol
@@ -8,9 +8,9 @@ pragma solidity 0.8.23;
  * @custom:security-contact bounty@safe.global
  */
 interface ISafeTokenLock {
-    event Locked(address indexed holder, uint96 amount);
-    event Unlocked(address indexed holder, uint32 indexed index, uint96 amount);
-    event Withdrawn(address indexed holder, uint32 indexed index, uint96 amount);
+    event Locked(address indexed holder, uint256 amount);
+    event Unlocked(address indexed holder, uint256 indexed index, uint256 amount);
+    event Withdrawn(address indexed holder, uint256 indexed index, uint256 amount);
 
     /**
      * @notice Error indicating an attempt to use zero tokens when locking or unlocking.
@@ -29,7 +29,7 @@ interface ISafeTokenLock {
      * Does not allow locking zero tokens.
      * Gas Usage (major): Token Transfer + SLOAD & SSTORE users[msg.sender] + Emit Event
      */
-    function lock(uint96 amount) external;
+    function lock(uint256 amount) external;
 
     /**
      * @notice Unlocks the specified amount of tokens.
@@ -39,7 +39,7 @@ interface ISafeTokenLock {
      * @dev Does not allow unlocking zero tokens.
      * Gas Usage (major): SLOAD & SSTORE users[msg.sender] + SLOAD COOLDOWN_PERIOD + SSTORE UnlockInfo + Emit Event
      */
-    function unlock(uint96 amount) external returns (uint32 index);
+    function unlock(uint256 amount) external returns (uint256 index);
 
     /**
      * @notice Withdraws the unlocked tokens of `maxUnlocks` oldest operations initiated by the caller.
@@ -50,12 +50,12 @@ interface ISafeTokenLock {
      * + n Zero assignment SSTORE unlocks[i][caller] + SSTORE users[caller] + SLOAD SAFE_TOKEN + Token Transfer
      * where n can be as high as max(`unlockEnd - unlockStart`, `maxUnlocks`).
      */
-    function withdraw(uint32 maxUnlocks) external returns (uint96 amount);
+    function withdraw(uint256 maxUnlocks) external returns (uint256 amount);
 
     /**
      * @notice Returns the amount of tokens associated to the specified holder.
      * @param holder The address of the holder.
      * @return amount The amount of (locked + to be unlocked + withdrawable) tokens of the holder.
      */
-    function totalBalance(address holder) external returns (uint96 amount);
+    function totalBalance(address holder) external returns (uint256 amount);
 }
